refactor: api 성격의 컴포넌트 패키지 이동
1. common.api.components 하위에 배치
   - resolver, interceptor, handler
