+++ src/main/java/com/cooper/concert/common/api/components/interceptor/QueueTokenValidationInterceptor.java
@@ -1,4 +1,4 @@
-package com.cooper.concert.interfaces.api.queues.interceptor;
+package com.cooper.concert.common.api.components.interceptor;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
