+++ src/main/java/com/cooper/concert/common/api/config/WebArgumentsConfig.java
@@ -8,7 +8,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import lombok.RequiredArgsConstructor;
 
-import com.cooper.concert.common.api.resolver.QueueTokenArgumentResolver;
+import com.cooper.concert.common.api.components.resolver.QueueTokenArgumentResolver;
 
 @Configuration
 @RequiredArgsConstructor
