+++ src/test/java/com/cooper/concert/interfaces/api/queues/controller/QueueTokenControllerTest.java
@@ -26,7 +26,7 @@ import com.cooper.concert.domain.queues.service.dto.QueueTokenIssueResult;
 import com.cooper.concert.domain.users.service.errors.UserErrorType;
 import com.cooper.concert.domain.users.service.errors.exception.UserNotFoundException;
 import com.cooper.concert.interfaces.api.queues.dto.request.QueueTokenIssueRequest;
-import com.cooper.concert.interfaces.api.queues.interceptor.QueueTokenValidationInterceptor;
+import com.cooper.concert.common.api.components.interceptor.QueueTokenValidationInterceptor;
 import com.cooper.concert.interfaces.api.queues.usecase.QueueTokenIssueUseCase;
 
 @WebMvcTest(value = QueueTokenController.class, excludeFilters = {@ComponentScan.Filter(
